resource_pack_client_response,protodef

+++ src/protocol/mcpe/packet/client_cache_status.rs
@@ -8,7 +8,7 @@ pub struct ClientCacheStatusPacket {
 }
 
 impl ClientCacheStatusPacket {
-    pub fn from_buf(buffer: Vec<u8>, offset: u64) -> Result<Self> {
+    pub fn from_buf(buffer: Vec<u8>, offset: usize) -> Result<Self> {
         let (enabled, _) = buffer.read_bool(offset)?;
         Ok(ClientCacheStatusPacket { enabled })
     }
